- Resolve conflict in MR !1846
implement JVET-S0219,
in order to implement extraction process that is decided to adopt by JVET_S0219, the equation (32) in the SPEC(JVET-S2001_vG) is needed to integrate. for this, JVET-R0193 and JVET-S0141 item47 is also implemented here.
please refer below meeting note for each aspect.
[JVET-S0219] Enable removal of parameter sets (i.e., PPS and APS) in non-output layer that is not used as reference for decoding of pictures in output layers during extraction process

AHG Recommendation (cleanup): Adopt APS aspect only.

[JVET-R0193] Signal max_tid_il_ref_pics_plus1 value separately for each direct reference layer of a layer, i.e. max_tid_il_ref_pics_plus1[ i ][ j ] for each direct reference layer j less than i, instead of single  max_tid_il_ref_pics_plus1[ i ]

AHG Recommendation (cleanup): Adopt. Text was provided in the contribution, and the authors are to supply the software

[JVET-S0141 item 47] In the general sub-bitstream extraction process, specify the conditions under which an output sub-bitstream is required to be a conforming bitstream such that the value of tIdTarget is specified to be in the range of 0 to vps_ptl_max_tid[ vps_ols_ptl_idx[ targetOlsIdx ] ], inclusive (instead of 0 to 6 inclusive)? (JVET-S0158 aspect 1)

Editor action item: The editor is suggested to include this change.