Fix #1353: Deblocking, ISP, and 4:2:2/4:4:4 chroma formats
When using ISP, the last TU in the CU contains the last TU for
luma but also the only TU for chroma. Max filter length is set
for chroma even when the luma TU is not aligned to 4x4 grid
(instead of being skipped, which is incorrect).
Although this fix changes setting of max filter length for the
4:2:0 chroma format as well, it doesn't change the deblocking
result in that case because the chroma TU size cannot be 8 or
larger when the last luma TU is not aligned to the 4x4 grid.